+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/extension/DrawableExt.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego.testUtils
+package co.omisego.omisego.extension
 
 /*
  * OmiseGO
