+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/model/BalanceTest.kt
@@ -12,7 +12,7 @@ import android.support.test.runner.AndroidJUnit4
 import co.omisego.omisego.extension.bd
 import co.omisego.omisego.helpers.delegation.GsonDelegator
 import co.omisego.omisego.helpers.delegation.ResourceFile
-import co.omisego.omisego.testUtils.validateParcel
+import co.omisego.omisego.utils.validateParcel
 import org.amshove.kluent.shouldBeInstanceOf
 import org.amshove.kluent.shouldEqual
 import org.amshove.kluent.shouldNotBe
