+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/model/DateTest.kt
@@ -10,7 +10,7 @@ package co.omisego.omisego.model
 import co.omisego.omisego.extension.toCalendar
 import co.omisego.omisego.helpers.delegation.GsonDelegator
 import co.omisego.omisego.helpers.delegation.ResourceFile
-import co.omisego.omisego.testUtils.DateConverter
+import co.omisego.omisego.utils.DateConverter
 import com.google.gson.reflect.TypeToken
 import org.amshove.kluent.shouldEqual
 import org.amshove.kluent.shouldNotBe
