+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/model/transaction/TransactionTest.kt
@@ -10,7 +10,7 @@ package co.omisego.omisego.model.transaction
 import co.omisego.omisego.helpers.delegation.GsonDelegator
 import co.omisego.omisego.helpers.delegation.ResourceFile
 import co.omisego.omisego.model.pagination.Paginable
-import co.omisego.omisego.testUtils.DateConverter
+import co.omisego.omisego.utils.DateConverter
 import org.amshove.kluent.shouldBeInstanceOf
 import org.amshove.kluent.shouldEqual
 import org.amshove.kluent.shouldEqualTo
