+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/network/ewallet/EWalletClientTest.kt
@@ -21,7 +21,7 @@ import co.omisego.omisego.model.Setting
 import co.omisego.omisego.model.User
 import co.omisego.omisego.model.WalletList
 import co.omisego.omisego.model.transaction.list.TransactionListParams
-import co.omisego.omisego.testUtils.Encryptor
+import co.omisego.omisego.utils.Encryptor
 import okhttp3.HttpUrl
 import okhttp3.mockwebserver.MockWebServer
 import org.amshove.kluent.mock
